feat: add the custom node
